+++ src/main/java/net/geforcemods/securitycraft/blocks/mines/BlockMine.java
@@ -83,6 +83,8 @@ public class BlockMine extends BlockExplosive {
 			return;
 		else if (entity instanceof EntityItem)
 			return;
+		else if (!state.getBoundingBox(world, pos).offset(pos).grow(0.01D).intersects(entity.getEntityBoundingBox()))
+			return;
 		else if (!EntityUtils.doesEntityOwn(entity, world, pos))
 			explode(world, pos);
 	}